Wired up the Event Aggregation
Wired up the Event Aggregation so when you log into the log in page the user is sent to the sales page. If a new request for the log in page is requested while the user is already logged on a new log on page will be requested so no user infomation would appear in the log in page.